+++ src/Constants/Discount.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace PlacetoPay\PaymentMethod\Constants;
+
+interface Discount
+{
+    public const UY_IVA_REFUND = '17934';
+
+    public const UY_IMESI_REFUND = '18083';
+
+    public const UY_FINANCIAL_INCLUSION = '19210';
+
+    public const UY_AFAM_REFUND = '18910';
+
+    public const UY_TAX_REFUND = '18999';
+
+    public const UY_NONE = '0';
+
+}