+++ src/Countries/CountryConfig.php
@@ -2,6 +2,7 @@
 
 namespace PlacetoPay\PaymentMethod\Countries;
 
+use PlacetoPay\PaymentMethod\Constants\Client;
 use PlacetoPay\PaymentMethod\Constants\Environment;
 
 abstract class CountryConfig implements CountryConfigInterface
@@ -11,12 +12,19 @@ abstract class CountryConfig implements CountryConfigInterface
         return true;
     }
 
-    public static function getEndpoints(): array
+    public static function getEndpoints(string $client): array
     {
         return [
             Environment::DEV => 'https://dev.placetopay.com/redirection',
             Environment::TEST => 'https://checkout-test.placetopay.com',
             Environment::PROD => 'https://checkout.placetopay.com',
+        ];
+    }
+
+    public static function getClients(): array
+    {
+        return [
+            unmaskString(Client::PTP) => __(unmaskString(Client::PTP), 'woocommerce-gateway-placetopay'),
         ];
     }
 }