+++ src/GatewayMethod.php
@@ -6,13 +6,16 @@ if (!defined('ABSPATH')) {
     exit; // Exit if accessed directly
 }
 
+use Dnetix\Redirection\Entities\PaymentModifier;
 use Dnetix\Redirection\Entities\Status;
 use Dnetix\Redirection\Entities\Transaction;
 use Dnetix\Redirection\Message\Notification;
 use Dnetix\Redirection\Message\RedirectInformation;
 use Dnetix\Redirection\PlacetoPay;
 use Exception;
+use PlacetoPay\PaymentMethod\Constants\Client;
 use PlacetoPay\PaymentMethod\Constants\Country;
+use PlacetoPay\PaymentMethod\Constants\Discount;
 use PlacetoPay\PaymentMethod\Constants\Environment;
 use PlacetoPay\PaymentMethod\Constants\Rules;
 use PlacetoPay\PaymentMethod\Countries\CountryConfigInterface;
@@ -54,7 +57,7 @@ class GatewayMethod extends WC_Payment_Gateway
      */
     const NOTIFICATION_RETURN_PAGE = 'placetopay_notification_return_page';
 
-    const EXPIRATION_TIME_MINUTES_LIMIT = 40320;
+    const EXPIRATION_TIME_MINUTES_LIMIT = 8640;
 
 
     /**
@@ -77,7 +80,6 @@ class GatewayMethod extends WC_Payment_Gateway
     private $expiration_time_minutes;
     private $endpoint;
     private $currency;
-    private $country;
     private $enviroment_mode;
     private $fill_buyer_information;
     private $login;
@@ -116,8 +118,8 @@ class GatewayMethod extends WC_Payment_Gateway
     public function configPaymentMethod()
     {
         $this->id = 'placetopay';
-        $this->title = $this->get_option('title');
-        $this->method_title = $this->getAppName();
+        $this->title = $this->get_option('client');
+        $this->method_title = $this->getClient();
         $this->method_description = __('Sells online safely and agile', 'woocommerce-gateway-placetopay');
         $this->has_fields = false;
 
@@ -128,9 +130,8 @@ class GatewayMethod extends WC_Payment_Gateway
         $this->endpoint = $this->settings['endpoint'];
         $this->expiration_time_minutes = $this->settings['expiration_time_minutes'];
         $this->fill_buyer_information = $this->get_option('fill_buyer_information');
-        $this->country = $this->get_option('country');
         $this->enviroment_mode = $this->get_option('enviroment_mode');
-        $this->description = $this->get_option('description');
+        $this->description = sprintf(__('Pay securely through %s.', 'woocommerce-gateway-placetopay'), $this->getClient());
         $this->login = $this->get_option('login');
         $this->tran_key = $this->get_option('tran_key');
         $this->redirect_page_id = $this->get_option('redirect_page_id');
@@ -200,12 +201,15 @@ class GatewayMethod extends WC_Payment_Gateway
 
         if (empty($url)) {
             // format: null
-            switch ($this->settings['country']) {
-                case Country::CL:
+            switch (unmaskString($this->getClient())) {
+                case Client::GNT:
                     $image = unmaskString('uggcf://onapb.fnagnaqre.py/hcybnqf/000/029/870/0620s532-9sp9-4248-o99r-78onr9s13r1q/bevtvany/Ybtb_JroPurpxbhg_Trgarg.fit');
                     break;
+                case Client::GOU:
+                    $image = unmaskString('uggcf://cynprgbcnl-fgngvp-cebq-ohpxrg.f3.hf-rnfg-2.nznmbanjf.pbz/tbhcntbf-pbz-pb/urnqre.fit');
+                    break;
                 default:
-                    $image = 'https://static.placetopay.com/placetopay-logo.svg';
+                    $image = unmaskString('uggcf://fgngvp.cynprgbcnl.pbz/cynprgbcnl-ybtb.fit');
             }
         } elseif ($this->checkValidUrl($url)) {
             // format: https://www.domain.test/image.svg
@@ -357,7 +361,6 @@ class GatewayMethod extends WC_Payment_Gateway
             ];
         }
 
-        $ref = $order->get_order_key() . '-' . time();
         $redirectUrl = $this->getRedirectUrl($order);
 
         //For wooCoomerce 2.0
@@ -394,7 +397,7 @@ class GatewayMethod extends WC_Payment_Gateway
             ],
             'payment' => [
                 'reference' => $orderNumber,
-                'description' => sprintf(__('Payment on %s No: %s', 'woocommerce-gateway-placetopay'), $this->title, $orderNumber),
+                'description' => sprintf(__('Payment on %s No: %s', 'woocommerce-gateway-placetopay'), $this->getClient(), $orderNumber),
                 'amount' => [
                     'currency' => $this->currency,
                     'total' => $order->get_total()
@@ -409,6 +412,22 @@ class GatewayMethod extends WC_Payment_Gateway
                 ],
             ],
         ];
+
+        if ($this->getCountry() === Country::UY) {
+            $discountCode = $this->get_option('discount');
+
+            if ($discountCode !== Discount::UY_NONE) {
+                $req['payment']['modifiers'] = [
+                    new PaymentModifier([
+                        'type' => PaymentModifier::TYPE_FEDERAL_GOVERNMENT,
+                        'code' => $discountCode,
+                        'additional' => [
+                            'invoice' => $this->get_option('invoice')
+                        ]
+                    ])
+                ];
+            }
+        }
 
         if (count($order->get_taxes()) > 0) {
             $req['payment']['amount']['taxes'] = $this->getOrderTaxes($order);
@@ -484,9 +503,9 @@ class GatewayMethod extends WC_Payment_Gateway
             $taxData = $tax->get_data();
 
             if (in_array($taxData['rate_id'], $valueAddedTaxType)) {
-                $shippingTax = (float) $order->get_shipping_tax();
+                $shippingTax = (float)$order->get_shipping_tax();
                 $totalTax = $shippingTax + $taxData['tax_total'];
-                $totalBase = $shippingTax > 0 ? ((float) $order->get_shipping_total() + $subTotal) : $subTotal;
+                $totalBase = $shippingTax > 0 ? ((float)$order->get_shipping_total() + $subTotal) : $subTotal;
 
                 $taxForP2P[] = [
                     'kind' => 'valueAddedTax',
@@ -543,7 +562,7 @@ class GatewayMethod extends WC_Payment_Gateway
             // Add information to the order to notify that exit to PlacetoPay
             // and invalidates the shopping cart
             $order = new WC_Order($orderId);
-            $order->update_status('on-hold', sprintf(__('Redirecting to %s', 'woocommerce-gateway-placetopay'), $this->getAppName()));
+            $order->update_status('on-hold', sprintf(__('Redirecting to %s', 'woocommerce-gateway-placetopay'), $this->getClient()));
 
             $this->resolveWebCheckout($order);
 
@@ -567,7 +586,7 @@ class GatewayMethod extends WC_Payment_Gateway
             return;
         }
 
-        wp_die(sprintf(__("%s Request Failure", 'woocommerce-gateway-placetopay'), $this->getAppName()));
+        wp_die(sprintf(__("%s Request Failure", 'woocommerce-gateway-placetopay'), $this->getClient()));
     }
 
     /**
@@ -615,7 +634,7 @@ class GatewayMethod extends WC_Payment_Gateway
         $currentPaymentStatus = get_post_meta($order->get_id(), self::META_STATUS, true);
 
         if ($isCallback && $currentPaymentStatus === Status::ST_APPROVED) {
-            $this->logger('Returning method is already '. $currentPaymentStatus, __METHOD__);
+            $this->logger('Returning method is already ' . $currentPaymentStatus, __METHOD__);
             return;
         }
 
@@ -710,7 +729,7 @@ class GatewayMethod extends WC_Payment_Gateway
                 if ($order->get_total() != floatval($totalAmount)) {
                     $message = sprintf(
                         __('Validation error: %s amounts do not match (gross %s).', 'woocommerce-gateway-placetopay'),
-                        $this->getAppName(),
+                        $this->getClient(),
                         $totalAmount
                     );
 
@@ -720,7 +739,7 @@ class GatewayMethod extends WC_Payment_Gateway
                 if (!empty($payerEmail)) {
                     update_post_meta(
                         $order->get_id(),
-                        sprintf(__('Payer %s email', 'woocommerce-gateway-placetopay'), $this->getAppName()),
+                        sprintf(__('Payer %s email', 'woocommerce-gateway-placetopay'), $this->getClient()),
                         $payerEmail
                     );
                 }
@@ -1161,7 +1180,7 @@ class GatewayMethod extends WC_Payment_Gateway
     /**
      * Get pages for return page setting
      *
-     * @param  boolean $title Title of the page
+     * @param boolean $title Title of the page
      * @return array
      */
     public function getPages($title = false)
@@ -1179,7 +1198,7 @@ class GatewayMethod extends WC_Payment_Gateway
         return $pageList;
     }
 
-    public function getCountryList()
+    public function getCountryList(): array
     {
         return [
             Country::BZ => __('Belize', 'woocommerce-gateway-placetopay'),
@@ -1194,12 +1213,33 @@ class GatewayMethod extends WC_Payment_Gateway
         ];
     }
 
-    /**
-     * Return list of environments for selection
-     *
-     * @return array
-     */
-    protected function getEnvironments()
+    public function getDiscounts(): array
+    {
+        return [
+            Discount::UY_NONE => __(Discount::UY_NONE, 'woocommerce-gateway-placetopay'),
+            Discount::UY_IVA_REFUND => __(Discount::UY_IVA_REFUND, 'woocommerce-gateway-placetopay'),
+            Discount::UY_IMESI_REFUND => __(Discount::UY_IMESI_REFUND, 'woocommerce-gateway-placetopay'),
+            Discount::UY_FINANCIAL_INCLUSION => __(Discount::UY_FINANCIAL_INCLUSION, 'woocommerce-gateway-placetopay'),
+            Discount::UY_AFAM_REFUND => __(Discount::UY_AFAM_REFUND, 'woocommerce-gateway-placetopay'),
+            Discount::UY_TAX_REFUND => __(Discount::UY_TAX_REFUND, 'woocommerce-gateway-placetopay'),
+        ];
+    }
+
+    public function getClientList(): array
+    {
+        /** @var CountryConfigInterface $config */
+        foreach (Country::COUNTRIES_CLIENT as $config) {
+            if (!$config::resolve($this->getCountry())) {
+                continue;
+            }
+
+            return $config::getClients();
+        }
+
+        return [];
+    }
+
+    public function getEnvironments(): array
     {
         return [
             Environment::DEV => __('Development', 'woocommerce-gateway-placetopay'),
@@ -1209,12 +1249,7 @@ class GatewayMethod extends WC_Payment_Gateway
         ];
     }
 
-    /**
-     * Get expiration time minutes list
-     *
-     * @return array
-     */
-    protected function getListOptionExpirationMinutes()
+    public function getListOptionExpirationMinutes(): array
     {
         $options = [];
         $format = '%d %s';
@@ -1227,33 +1262,16 @@ class GatewayMethod extends WC_Payment_Gateway
             } elseif ($minutes >= 60 && $minutes < 1440) {
                 $options[$minutes] = sprintf($format, $minutes / 60, __('Hour(s)', 'woocommerce-gateway-placetopay'));
                 $minutes += 60;
-            } elseif ($minutes >= 1440 && $minutes < 10080) {
+            } else {
                 $options[$minutes] = sprintf($format, $minutes / 1440, __('Day(s)', 'woocommerce-gateway-placetopay'));
                 $minutes += 1440;
-            } elseif ($minutes >= 10080 && $minutes < 40320) {
-                $options[$minutes] = sprintf(
-                    $format,
-                    $minutes / 10080,
-                    __('Week(s)', 'woocommerce-gateway-placetopay')
-                );
-                $minutes += 10080;
-            } else {
-                $options[$minutes] = sprintf(
-                    $format,
-                    $minutes / 40320,
-                    __('Month(s)', 'woocommerce-gateway-placetopay')
-                );
-                $minutes += 40320;
             }
         }
 
         return $options;
     }
 
-    /**
-     * @return array
-     */
-    protected function getListTaxes()
+    public function getListTaxes(): array
     {
         $countries = $this->getCountryList();
         $formatTaxItem = '%s( %s ) - %s - %s %%';
@@ -1276,17 +1294,17 @@ class GatewayMethod extends WC_Payment_Gateway
         return $taxList;
     }
 
-    protected function getDefaultAppName(): string
+    public function getDefaultClient(): string
     {
-        return $this->getWooCommerceCountry() === Country::CL ? unmaskString('Trgarg') : 'PlacetoPay';
+        return $this->getCountry() === Country::CL ? unmaskString('Trgarg') : 'PlacetoPay';
     }
 
-    public function getAppName(): string
+    public function getClient(): string
     {
-        return !empty($this->title) ? $this->title : $this->getDefaultAppName() ;
+        return !empty($this->title) ? $this->title : $this->getDefaultClient();
     }
 
-    protected function getWooCommerceCountry(): string
+    public function getCountry(): string
     {
         return explode(':', get_option('woocommerce_default_country'))[0];
     }
@@ -1296,7 +1314,7 @@ class GatewayMethod extends WC_Payment_Gateway
         $domain = $_SERVER['HTTP_HOST'] ?? ($_SERVER['SERVER_NAME'] ?? 'localhost');
 
         return [
-            'User-Agent' => "woocommerce-gateway-placetopay/{$this->version} (origin:$domain; vr:" . WOOCOMMERCE_VERSION. ')',
+            'User-Agent' => "woocommerce-gateway-placetopay/{$this->version} (origin:$domain; vr:" . WOOCOMMERCE_VERSION . ')',
             'X-Source-Platform' => 'woocommerce',
         ];
     }
@@ -1324,7 +1342,7 @@ class GatewayMethod extends WC_Payment_Gateway
     /**
      * Get the class name with namespaces modificated
      *
-     * @param  boolean $lowercase
+     * @param boolean $lowercase
      * @return string
      */
     private function getClassName($lowercase = false)
@@ -1411,10 +1429,10 @@ class GatewayMethod extends WC_Payment_Gateway
     private function wooCommerceVersionCompare($version, $operator = '>=')
     {
         return defined('WOOCOMMERCE_VERSION') && version_compare(
-            WOOCOMMERCE_VERSION,
-            $version,
-            $operator
-        );
+                WOOCOMMERCE_VERSION,
+                $version,
+                $operator
+            );
     }
 
     private function configureEnvironment()
@@ -1484,11 +1502,6 @@ class GatewayMethod extends WC_Payment_Gateway
         return null;
     }
 
-    /**
-     * @param string $version
-     *
-     * @return bool
-     */
     public static function versionCheck(string $version = '3.0'): bool
     {
         if (class_exists('WooCommerce')) {
@@ -1507,11 +1520,11 @@ class GatewayMethod extends WC_Payment_Gateway
         $installments = ['installments', 'installment'];
 
         if (isset($additionalData['credit'])) {
-            $key = array_keys(array_filter($additionalData['credit'], function($item) use ($installments) {
+            $key = array_keys(array_filter($additionalData['credit'], function ($item) use ($installments) {
                 return in_array($item, $installments);
             }, ARRAY_FILTER_USE_KEY));
 
-            if(isset($key[0])) {
+            if (isset($key[0])) {
                 return $additionalData['credit'][$key[0]];
             }
         }
@@ -1523,11 +1536,11 @@ class GatewayMethod extends WC_Payment_Gateway
     {
         /** @var CountryConfigInterface $config */
         foreach (Country::COUNTRIES_CONFIG as $config) {
-            if (!$config::resolve($this->settings['country'])) {
+            if (!$config::resolve($this->getCountry())) {
                 continue;
             }
 
-            return $config::getEndpoints();
+            return $config::getEndpoints($this->getClient());
         }
 
         return [];
@@ -1581,7 +1594,7 @@ class GatewayMethod extends WC_Payment_Gateway
                 message: "' . esc_js(sprintf(__(
                 'We are now redirecting you to %s to make payment, if you are not redirected please press the bottom.',
                 'woocommerce-gateway-placetopay'
-            ), $this->getAppName())) . '",
+            ), $this->getClient())) . '",
                 baseZ: 99999,
                 overlayCSS: { background: "#fff", opacity: 0.6 },
                 css: {
@@ -1615,8 +1628,8 @@ class GatewayMethod extends WC_Payment_Gateway
         $redirectUrl = $this->getRedirectUrl($order);
 
         $redirectUrl = add_query_arg([
-            'wc-api'=> $this->getClassName(),
-            'order_id'=> $order->get_id(),
+            'wc-api' => $this->getClassName(),
+            'order_id' => $order->get_id(),
         ], $redirectUrl);
 
         return $redirectUrl . '&key=' . $order->get_order_key() . '-' . time();